Let android-entry run as system user
For security reason, our customer want all the processes of
a container to run as non-root user, so let android-entry
run as system user.
We add a '-y' flag to control this behavior, if we pass '-y' to
android container, then the rootless feature will be enabled.
or root processes will still run as root.

Tracked-On: OAM-90885
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>